# perf(utils/url): improve performance of url utility