@EdamAme-x 2024-10-31
Benchmark Source
```ts
import { run, bench } from 'mitata'

const urls = ['', 'hello', 'hello world', 'hono+is+cool', 'hono+is+very+hot']

const regex = (value: string) => {
  return /\+/.test(value)
}

const indexOf = (value: string) => {
  return value.indexOf('+') !== -1
}

bench('Regex', () => {
  for (const url of urls) {
    regex(url)
  }
})

bench('IndexOf', () => {
  for (const url of urls) {
    indexOf(url)
  }
})

await run()
```