@usualoma 2024-10-31
@EdamAme-x Thank you!

I think you're right about `indexOf`.

Do you think the following will actually be effective?
https://github.com/honojs/hono/pull/3593/files#diff-aab4ceb6d3254a7b948f95d8057cd8d91cd43c09f95569885cbce6d8b079d926L107-R107

I don't think there is any difference in major runtimes, and I think it is unlikely that there will be any difference in the future.

### benchmark

```ts
import { run, bench, group } from 'mitata'

bench('noop', () => {})

const url = 'https://example.com/path?query=string#fragment'

group('for loop', () => {
  bench('use url.length', () => {
    let i = 0
    for (; i < url.length; i++) {
      url.charCodeAt(i)
    }
  })

  bench('use cached length', () => {
    let i = 0
    for (const len = url.length; i < len; i++) {
      url.charCodeAt(i)
    }
  })
})

await run()
```

```
cpu: Apple M2 Pro
runtime: node v22.2.0 (arm64-darwin)

benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
noop                   73 ps/iter      (20 ps … 76.35 ns)     82 ps    102 ps    203 ps !

• for loop
--------------------------------------------------------- -----------------------------
use url.length       13.9 ns/iter     (13.71 ns … 103 ns)  13.79 ns  16.07 ns  22.28 ns
use cached length   13.97 ns/iter    (13.71 ns … 69.4 ns)  13.83 ns  16.26 ns   59.9 ns

summary for for loop
  use url.length
   1x faster than use cached length
```

```
cpu: Apple M2 Pro
runtime: bun 1.1.30 (arm64-darwin)

benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
noop                   44 ps/iter       (0 ps … 30.58 ns)     41 ps     82 ps    102 ps !

• for loop
--------------------------------------------------------- -----------------------------
use url.length      13.87 ns/iter   (13.43 ns … 52.49 ns)  13.71 ns  17.56 ns  19.47 ns
use cached length   13.95 ns/iter   (13.41 ns … 60.97 ns)  13.98 ns  17.99 ns  21.99 ns

summary for for loop
  use url.length
   1.01x faster than use cached length
```

```
cpu: Apple M2 Pro
runtime: deno 2.0.0 (aarch64-apple-darwin)

benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
noop                  508 ps/iter     (427 ps … 36.56 ns)    509 ps    631 ps    753 ps

• for loop
--------------------------------------------------------- -----------------------------
use url.length       8.96 ns/iter    (8.73 ns … 78.59 ns)   8.83 ns  11.31 ns  19.69 ns
use cached length    8.95 ns/iter    (8.71 ns … 75.87 ns)   8.81 ns  11.21 ns  62.22 ns

summary for for loop
  use cached length
   1x faster than use url.length
```